move all functions without return or redirect to helper
