+++ server/apps/main/views.py
@@ -2,21 +2,29 @@ import datetime
 import io
 import json
 import logging
-import uuid
 
 import requests
 from django.conf import settings
 from django.contrib.auth import login, logout
-from django.forms.models import model_to_dict
 from django.shortcuts import redirect, render
 from openhumans.models import OpenHumansMember
 from django.db.models import Q
-from django.contrib.auth.models import Group
-
 
 from .models import PublicExperience, ExperienceHistory
 
 from .forms import ShareExperienceForm, ModerateExperienceForm
+
+from .helpers import (is_moderator,
+                        model_to_form,
+                        extract_experience_details,
+                        reformat_date_string,
+                        get_review_status,
+                        get_oh_file,
+                        delete_single_file,
+                        upload,
+                        make_uuid,
+                        delete_PE,
+                        update_public_experience_db)
 
 logger = logging.getLogger(__name__)
 
@@ -110,120 +118,120 @@ def view_experience(request, uuid):
         redirect('index')
 
 
-def update_public_experience_db(data, uuid, ohmember, change_type="Edit", change_comments=None):
-    """Updates the public experience database for the given uuid.
+# def update_public_experience_db(data, uuid, ohmember, change_type="Edit", change_comments=None):
+#     """Updates the public experience database for the given uuid.
 
-    If data is tagged as viewable, an experience will be updated or inserted.
-    If a data is tagged as not public, this function ensures that it is absent from the pe db.
+#     If data is tagged as viewable, an experience will be updated or inserted.
+#     If a data is tagged as not public, this function ensures that it is absent from the pe db.
 
-    Args:
-        data (dict): an experience
-        uuid (str): unique identifier
-        ohmember : request.user.openhumansmember
-        moderation_status (str, optional): Defaults to 'in review'.
-    """
+#     Args:
+#         data (dict): an experience
+#         uuid (str): unique identifier
+#         ohmember : request.user.openhumansmember
+#         moderation_status (str, optional): Defaults to 'in review'.
+#     """
 
-    if data.pop('viewable',False):
-        data.pop('open_humans_member', None) # Remove from dict as needs to be an object not string
-        data.pop('experience_id', None) # Remove from dict as needs to be an object not string
-        data.pop('moderation_comments', None) # Only need when moderating
+#     if data.pop('viewable',False):
+#         data.pop('open_humans_member', None) # Remove from dict as needs to be an object not string
+#         data.pop('experience_id', None) # Remove from dict as needs to be an object not string
+#         data.pop('moderation_comments', None) # Only need when moderating
         
-        pe = PublicExperience(open_humans_member=ohmember, experience_id=uuid, **data)
+#         pe = PublicExperience(open_humans_member=ohmember, experience_id=uuid, **data)
         
-        # .save() updates if primary key exists, inserts otherwise. 
-        pe.save()
+#         # .save() updates if primary key exists, inserts otherwise. 
+#         pe.save()
 
-        if pe.experiencehistory_set.count() == 0:
-            change_type = "Make Public"
-            change_comments = "Story flagged to be shared on AutSPACE website"
-        elif change_comments == None:
-            change_comments = "Story edited"
-        # Produce and add the ExperienceHistory object to the public experience
-        eh = ExperienceHistory(experience=pe,
-                                change_type = change_type, 
-                                changed_at = datetime.datetime.now(),
-                                changed_by = str(ohmember),
-                                change_comments = change_comments)
+#         if pe.experiencehistory_set.count() == 0:
+#             change_type = "Make Public"
+#             change_comments = "Story flagged to be shared on AutSPACE website"
+#         elif change_comments == None:
+#             change_comments = "Story edited"
+#         # Produce and add the ExperienceHistory object to the public experience
+#         eh = ExperienceHistory(experience=pe,
+#                                 change_type = change_type, 
+#                                 changed_at = datetime.datetime.now(),
+#                                 changed_by = str(ohmember),
+#                                 change_comments = change_comments)
 
-        eh.save()
-        # change_type = models.TextField()
-        # changed_at = models.DateTimeField(auto_now=True)
-        # changed_by = models.TextField()
-        # change_comments = models.TextField(default="")  
+#         eh.save()
+#         # change_type = models.TextField()
+#         # changed_at = models.DateTimeField(auto_now=True)
+#         # changed_by = models.TextField()
+#         # change_comments = models.TextField(default="")  
         
-    else:
-        delete_PE(uuid,ohmember)
+#     else:
+#         delete_PE(uuid,ohmember)
 
 
-def delete_PE(uuid, ohmember):
-    if PublicExperience.objects.filter(experience_id=uuid, open_humans_member=ohmember).exists():
-            PublicExperience.objects.get(experience_id=uuid, open_humans_member=ohmember).delete()
+# def delete_PE(uuid, ohmember):
+#     if PublicExperience.objects.filter(experience_id=uuid, open_humans_member=ohmember).exists():
+#             PublicExperience.objects.get(experience_id=uuid, open_humans_member=ohmember).delete()
 
 
-def make_uuid():
-    return str(uuid.uuid1())
+# def make_uuid():
+#     return str(uuid.uuid1())
 
-def upload(data, uuid, ohmember):
-    """Uploads a dictionary representation of an experience to open humans.
+# def upload(data, uuid, ohmember):
+#     """Uploads a dictionary representation of an experience to open humans.
 
-    Args:
-        data (dict): an experience
-        uuid (str): unique identifier
-        ohmember : request.user.openhumansmember
-    """
+#     Args:
+#         data (dict): an experience
+#         uuid (str): unique identifier
+#         ohmember : request.user.openhumansmember
+#     """
 
-    output_json = {
-            'data': data,
-            'timestamp': str(datetime.datetime.now())}
+#     output_json = {
+#             'data': data,
+#             'timestamp': str(datetime.datetime.now())}
 
-    # by saving the output json into metadata we can access the fields easily through request.user.openhumansmember.list_files().
-    metadata = {
+#     # by saving the output json into metadata we can access the fields easily through request.user.openhumansmember.list_files().
+#     metadata = {
 
-        'uuid': uuid,   
-        'description': data.get('title_text'),
-        'tags': make_tags(data),
-        **output_json,
-        }
+#         'uuid': uuid,   
+#         'description': data.get('title_text'),
+#         'tags': make_tags(data),
+#         **output_json,
+#         }
 
-    # create stream for oh upload
-    output = io.StringIO()
-    output.write(json.dumps(output_json))
-    output.seek(0)
+#     # create stream for oh upload
+#     output = io.StringIO()
+#     output.write(json.dumps(output_json))
+#     output.seek(0)
 
 
-    ohmember.upload(
-        stream=output,
-        filename=f"{uuid}.json", #filename is Autspaces_timestamp
-        metadata=metadata)
+#     ohmember.upload(
+#         stream=output,
+#         filename=f"{uuid}.json", #filename is Autspaces_timestamp
+#         metadata=metadata)
 
-def make_tags(data):
-    """builds list of tags based on data"""
+# def make_tags(data):
+#     """builds list of tags based on data"""
 
-    tag_map = {'viewable': {'True':'public',
-                            'False':'not public'},
-               'research': {'True':'research',
-                            'False':'non-research'},
-                'drug':    {'True': 'drugs',
-                            'False': ''},
-                'abuse':    {'True': 'abuse',
-                            'False': ''},
-                'negbody':  {'True': 'negative body',
-                            'False': ''},
-                'violence': {'True': 'violence',
-                            'False': ''},
-                'mentalhealth': {'True': 'mental health',
-                            'False': ''},
-                'moderation_status': {'True': '',
-                            'False': 'in review'}
-                            }
+#     tag_map = {'viewable': {'True':'public',
+#                             'False':'not public'},
+#                'research': {'True':'research',
+#                             'False':'non-research'},
+#                 'drug':    {'True': 'drugs',
+#                             'False': ''},
+#                 'abuse':    {'True': 'abuse',
+#                             'False': ''},
+#                 'negbody':  {'True': 'negative body',
+#                             'False': ''},
+#                 'violence': {'True': 'violence',
+#                             'False': ''},
+#                 'mentalhealth': {'True': 'mental health',
+#                             'False': ''},
+#                 'moderation_status': {'True': '',
+#                             'False': 'in review'}
+#                             }
 
-    tags = [tag_map[k].get(str(v))
-            for k,v in data.items()
-            if k in tag_map.keys()]
-    if data["other"] != '':
-        tags.append("Other triggering label")
+#     tags = [tag_map[k].get(str(v))
+#             for k,v in data.items()
+#             if k in tag_map.keys()]
+#     if data["other"] != '':
+#         tags.append("Other triggering label")
 
-    return tags
+#     return tags
 
 def delete_experience(request, uuid, title):
     # TODO: we currently are passing title via url because it is nice to display it in the confirmation. We could improve the deletion process by having a javascript layover.
@@ -245,40 +253,40 @@ def delete_experience(request, uuid, title):
         return redirect('index')
 
 
-def delete_single_file(uuid, ohmember):
-    """Deletes a given file id and uuid from openhumans and ensures absence from local PublicExperiences database.
+# def delete_single_file(uuid, ohmember):
+#     """Deletes a given file id and uuid from openhumans and ensures absence from local PublicExperiences database.
 
-    Args:
+#     Args:
 
-        uuid (str): unique identifier, used for PublicExperience primary key and for OpenHumans filename.
-        ohmember : request.user.openhumansmember
-    """
+#         uuid (str): unique identifier, used for PublicExperience primary key and for OpenHumans filename.
+#         ohmember : request.user.openhumansmember
+#     """
 
-    ohmember.delete_single_file(file_basename=f"{uuid}.json")
-    delete_PE(uuid,ohmember)
+#     ohmember.delete_single_file(file_basename=f"{uuid}.json")
+#     delete_PE(uuid,ohmember)
 
-def get_oh_file(ohmember, uuid):
-    """Returns a single file from OpenHumans, filtered by uuid.
+# def get_oh_file(ohmember, uuid):
+#     """Returns a single file from OpenHumans, filtered by uuid.
 
-    Args:
-        ohmember : request.user.openhumansmember
-        uuid (str): unique identifier
+#     Args:
+#         ohmember : request.user.openhumansmember
+#         uuid (str): unique identifier
 
-    Raises:
-        Exception: If uuid belongs to more than one file.
+#     Raises:
+#         Exception: If uuid belongs to more than one file.
 
-    Returns:
-        file (dict): dictionary representation of OpenHumans file. Returns None if uuid is not matched.
-    """
-    files = ohmember.list_files()
-    file = [f for f in files if f["metadata"]["uuid"]==uuid]
+#     Returns:
+#         file (dict): dictionary representation of OpenHumans file. Returns None if uuid is not matched.
+#     """
+#     files = ohmember.list_files()
+#     file = [f for f in files if f["metadata"]["uuid"]==uuid]
 
-    if len(file)==0:
-        file=None
-    elif len(file)>1:
-        raise Exception("duplicate uuids in open humans")
+#     if len(file)==0:
+#         file=None
+#     elif len(file)>1:
+#         raise Exception("duplicate uuids in open humans")
 
-    return file[0]
+#     return file[0]
 
 
 
@@ -456,33 +464,33 @@ def registration(request):
 def signup_frame4_test(request):
     return render(request, "main/signup1.html")
 
-def get_review_status(files):
-    """Given a list of files, count the number of each moderation status of the publicly viewable files
+# def get_review_status(files):
+#     """Given a list of files, count the number of each moderation status of the publicly viewable files
 
-    Args:
-        files (dict): list of files, which are dictionaries. See `upload()`.
+#     Args:
+#         files (dict): list of files, which are dictionaries. See `upload()`.
 
-    Returns:
-        statuses (dict): counts of moderation statuses
-    """
+#     Returns:
+#         statuses (dict): counts of moderation statuses
+#     """
 
 
-    status_list = [f['metadata']['data']['moderation_status'].replace(' ','_') for f in files if f['metadata']['data']['viewable']]
+#     status_list = [f['metadata']['data']['moderation_status'].replace(' ','_') for f in files if f['metadata']['data']['viewable']]
 
-    statuses = {f"n_{s}":status_list.count(s) for s in set(status_list)}
+#     statuses = {f"n_{s}":status_list.count(s) for s in set(status_list)}
 
-    statuses["n_viewable"] = len(status_list)
-    statuses["n_moderated"] = status_list.count('approved') + status_list.count('rejected')
+#     statuses["n_viewable"] = len(status_list)
+#     statuses["n_moderated"] = status_list.count('approved') + status_list.count('rejected')
 
-    return statuses
+#     return statuses
 
-def reformat_date_string(context):
-     """
-     Convert the date string back to a datetime object
-     """
-     for file in context['files']:
-         file['created'] = datetime.datetime.strptime(file['created'], '%Y-%m-%dT%H:%M:%S.%fZ')
-     return context
+# def reformat_date_string(context):
+#      """
+#      Convert the date string back to a datetime object
+#      """
+#      for file in context['files']:
+#          file['created'] = datetime.datetime.strptime(file['created'], '%Y-%m-%dT%H:%M:%S.%fZ')
+#      return context
 
 def my_stories(request):
     if request.user.is_authenticated:
@@ -505,18 +513,18 @@ def confirmation_page(request):
 def about_us(request):
     return render(request, "main/about_us.html")
 
-# def what_autism_is(request):
-#     auth_url = OpenHumansMember.get_auth_url()
-#     context = {'auth_url': auth_url,
-#                'oh_proj_page': settings.OH_PROJ_PAGE}
-#     if request.user.is_authenticated:
-#         return redirect('main:what_autism_is')
-#     return render(request, 'main/what_autism_is.html', context=context)
+## def what_autism_is(request):
+##     auth_url = OpenHumansMember.get_auth_url()
+##     context = {'auth_url': auth_url,
+##                'oh_proj_page': settings.OH_PROJ_PAGE}
+##     if request.user.is_authenticated:
+##         return redirect('main:what_autism_is')
+##     return render(request, 'main/what_autism_is.html', context=context)
 def navigation(request):
     return render(request, "main/navigation.html")
 
-# def about_us(request):
-#     return render(request, "main/about_us.html")
+## def about_us(request):
+##     return render(request, "main/about_us.html")
 
 def what_autism_is(request):
     return render(request, "main/what_autism_is.html")
@@ -524,33 +532,20 @@ def what_autism_is(request):
 def footer(request):
     return render(request, "main/footer.html")
 
-def process_trigger_warnings(form):
-    """
-    Get the mutable trigger warnings from the data
-    """
-    form.is_valid()
-    # Populate with values from moderation form
-    trigger_warning_dict = {k: form.cleaned_data[k] for k in form.data.keys() if k not in ['csrfmiddlewaretoken', 'research']}
+# def process_trigger_warnings(form):
+#     """
+#     Get the mutable trigger warnings from the data
+#     """
+#     form.is_valid()
+#     # Populate with values from moderation form
+#     trigger_warning_dict = {k: form.cleaned_data[k] for k in form.data.keys() if k not in ['csrfmiddlewaretoken', 'research']}
 
-    # Set all others as False
-    for key in ['abuse', 'violence', 'drug', 'mentalhealth', 'negbody']:
-        if key not in trigger_warning_dict.keys():
-            trigger_warning_dict[key] = False
+#     # Set all others as False
+#     for key in ['abuse', 'violence', 'drug', 'mentalhealth', 'negbody']:
+#         if key not in trigger_warning_dict.keys():
+#             trigger_warning_dict[key] = False
     
-    return trigger_warning_dict
-    
-
-def extract_experience_details(model):
-    """
-    Extract the immutable details of the story and the sharing options only
-    """
-    model_dict = model_to_dict(model)
-
-    for key in ['abuse', 'violence', 'drug', 'mentalhealth', 'negbody', 'other', 'moderation_status']:
-        model_dict.pop(key, None)
-    model_dict['viewable'] = True # Only moderate viewable experiences
-
-    return model_dict
+#     return trigger_warning_dict
 
 def moderate_experience(request, uuid):
     if request.user.is_authenticated and is_moderator(request.user):
@@ -602,22 +597,22 @@ def moderate_experience(request, uuid):
     else:
         redirect('index')
 
-def model_to_form(model, disable_moderator = False):
-    model_dict = model_to_dict(model)
+# def model_to_form(model, disable_moderator = False):
+#     model_dict = model_to_dict(model)
 
-    form = ModerateExperienceForm({**model_dict,
-        "viewable":True #we only moderate public experiences
-    }, disable_moderator=disable_moderator)
+#     form = ModerateExperienceForm({**model_dict,
+#         "viewable":True #we only moderate public experiences
+#     }, disable_moderator=disable_moderator)
 
-    return form
+#     return form
 
 
-def is_moderator(user):
-  """return membership of moderator group"""
+# def is_moderator(user):
+#   """return membership of moderator group"""
 
-  try:
-    group = Group.objects.get(user=user)
-    return (group.name == "Moderators")
+#   try:
+#     group = Group.objects.get(user=user)
+#     return (group.name == "Moderators")
 
-  except Group.DoesNotExist:
-    return False
+#   except Group.DoesNotExist:
+#     return False
